+++ core/blitzkrieg_core/src/lib.rs
@@ -42,6 +42,7 @@ pub mod signal;
 pub mod sim;
 pub mod strategies;
 pub mod strategy_engine;
+pub mod strategy_state;
 pub mod trade_db;
 
 pub const CORE_VERSION: &str = env!("CARGO_PKG_VERSION");